Remove obsolete Utility class from winform project
